Move generate options to generate subcmd
